Provide decomposed version to indirectly set packed version
The packed version of NSIS is used for feature detection in scripts.
If VER_MAJOR and VER_MINOR are not set then VER_PACKED is going
to be assigned the default value 0x03003666. VER_PACKED is set to
a value reflecting the version by providing the build parameters
VER_MAJOR, VER_MINOR and VER_REVISION.